requirements.txt
flask==3.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
redis==5.0.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
